Created first migration

+++ AirLineAPI/Model/PassengerTimeTable.cs
@@ -7,9 +7,9 @@ namespace AirLineAPI.Model
 {
     public class PassengerTimeTable
     {
-        public int PassengerId { get; set; }
+        public long PassengerId { get; set; }
         public Passenger Passenger { get; set; }
-        public int TimeTableId { get; set; }
+        public long TimeTableId { get; set; }
         public TimeTable TimeTable { get; set; }
     }
 }
